Hide tooltip for null content
Sometimes one would like to hide the tooltip for some elements.  I think the most logical way to do this is to not show tooltips if the content is null.  This mimics d3's behavior of deleting html content when null is passed to [`selection.html()`](https://github.com/mbostock/d3/wiki/Selections#html).

An example of when this would be useful is the following code snippet, where you don't want to show tooltips for empty boxes:
````
tip.html(function(d){
  if(d.contents) {
    return d.contents.join("<br>");
  } else {
    return null;
  }
});
````